Add failure cases to `Transactions::Steps::DeleteWorkStep`
When the work is not created or already destroyed, we want to fail this step.

+++ app/transactions/hyrax/transactions/steps/delete_work_step.rb
@@ -7,6 +7,9 @@ module Hyrax
         include Dry::Transaction::Operation
 
         def call(work)
+          return Failure(:not_created) unless work.persisted?
+          return Failure(:already_destroyed) if work.destroyed?
+
           work.destroy
           Success(work)
         end
